main: move io_read_write to lib
